- Fixed image_write_rgb() problems

+++ inc/image.h
@@ -6,8 +6,8 @@
 #include <fstream>
 #include <string>
 
-#include <inc/v3.h>
+#include "./v3.h"
 
-int image_write_rgb(std::string filename, v3 **img, uint32_t img_height, uint32_t img_width, uint32_t img_maxval = 255);
+int image_write_rgb(std::string filename, v3 **img, uint32_t img_height, uint32_t img_width, uint8_t img_maxval = 255);
 
 #endif
